requirements.txt
streamlit==1.38.0
ultralytics==8.2.55
torch==2.2.2
torchvision==0.17.2
pillow==10.2.0
plotly==5.22.0
numpy==1.26.4
